Modularise schema
